# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/service-requests/": {
            "get": {
                "tags": [
                    "service_requests"
                ],
                "summary": "Read Service Requests",
                "operationId": "read_service_requests_api_v1_service_requests__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListServiceRequestsResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "service_requests"
                ],
                "summary": "Create New Service Request",
                "operationId": "create_new_service_request_api_v1_service_requests__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Request"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateServiceRequestResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/service-requests/{request_id}/status": {
            "put": {
                "tags": [
                    "service_requests"
                ],
                "summary": "Modify Service Request Status",
                "operationId": "modify_service_request_status_api_v1_service_requests__request_id__status_put",
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Request Id"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateServiceRequestStatusResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/services/": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Read Services",
                "operationId": "read_services_api_v1_services__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListServicesResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "services"
                ],
                "summary": "Create New Service",
                "operationId": "create_new_service_api_v1_services__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Service"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateServiceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/services/{service_id}/quantity": {
            "put": {
                "tags": [
                    "services"
                ],
                "summary": "Modify Service Quantity",
                "operationId": "modify_service_quantity_api_v1_services__service_id__quantity_put",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Service Id"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Quantity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateServiceQuantityResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateServiceRequestResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ServiceRequestResponse"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "CreateServiceRequestResponse"
            },
            "CreateServiceResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Service"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "CreateServiceResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ListServiceRequestsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/ServiceRequestResponse"
                        },
                        "type": "array",
                        "title": "Data"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "ListServiceRequestsResponse"
            },
            "ListServicesResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/Service"
                        },
                        "type": "array",
                        "title": "Data"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "ListServicesResponse"
            },
            "Service": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "quantity": {
                        "type": "integer",
                        "title": "Quantity"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "quantity",
                    "created_at"
                ],
                "title": "Service"
            },
            "ServiceRequestResponse": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "service_id": {
                        "type": "integer",
                        "title": "Service Id"
                    },
                    "requester_name": {
                        "type": "string",
                        "title": "Requester Name"
                    },
                    "request_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Request Date"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ServiceStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "service_id",
                    "requester_name",
                    "request_date",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "title": "ServiceRequestResponse"
            },
            "ServiceStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "approved",
                    "returned"
                ],
                "title": "ServiceStatus"
            },
            "UpdateServiceQuantityResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Service"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message"
                ],
                "title": "UpdateServiceQuantityResponse"
            },
            "UpdateServiceRequestStatusResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ServiceRequestResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "UpdateServiceRequestStatusResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}